test(qa): keep retained-final harness in its owning test
